requirements.txt
pandas
spotipy
google-cloud-bigquery
pyarrow